Add feature priority (#7)
Although a setting-like argument to the parser, it isn't part of the 
settings and is therefore not borne by the settings list nor accessed 
via the Feature get_setting_value() instance method.

Instead, a utility getter is provided that sorts the features in order 
of priority, for both name sections and their application on the input.