Generalize alpha to a cross-feature setting (#12)
This unfortunately means cluttering up the Feature class even further 
with everything necessary to apply it. That class is starting to become 
a mess and might need to be fragmented.

Misc : since this isn't a setting similar to the others, they have been 
grouped in a list called enable_settings and it is in a separate list, 
video_settings. The containing file has been renamed to reflect this.