Introduce VideoInfo type
This allows for the Feature __call__() method to always have access to 
this info even if the feature filter does not require any of it. This 
will come in handy next commit.

On the purely cosmetic side, it also makes it unnecessary to use the 
locals() dict in the main script, which is better practice.

supplemental_arguments have also been renamed to 
required_video_information, which is more in line with what they are.